single test acquisition with comedi counter and dac
git-svn-id: https://svn.code.sf.net/p/opticsbenchui/code@35 19f4beda-a93a-4e52-9a4f-25590d2c5071

+++ AcquisitionSequence.h
@@ -75,7 +75,7 @@ class AcquisitionSequence
   float   position;
 
   // Dac attributes
-  float   dacValue;
+  double   dacValue;
   int     dacOutput;
 
   // Counter Comedi attributes
