+++ Dac.h
@@ -30,7 +30,7 @@ class Dac : public QObject
 
   virtual bool connectDac(QString newdac) = 0;
   virtual bool resetDac(QString newdac) = 0;
-  virtual bool setDacValue(QString newdac, int output, float value) = 0;
+  virtual bool setDacValue(QString newdac, int output, double value) = 0;
   virtual bool updateDBValues(QString newdac) = 0;
   virtual void setDbPath(QString _path) = 0;
